Clean: move fncs from main to seperate files
